refactor: 스텝 리스트 추가 및 로직 개선
