+++ src/navigators/StackNavigator.tsx
@@ -3,11 +3,8 @@ import {useRecoilValue} from 'recoil';
 import {stepsState} from 'libs/recoil/steps';
 import TabNavigator from 'navigators/TabNavigator';
 import {StackMenu} from 'navigators/constants/menu';
-import {Steps, steps} from 'navigators/constants/step';
+import {stepInfoList, steps} from 'navigators/constants/steps';
 import {StackParamList} from 'navigators/types';
-import Description from 'screens/OnBoarding/Description';
-import Login from 'screens/OnBoarding/Login';
-import UserInfo from 'screens/OnBoarding/UserInfo';
 import Funnel from 'components/@common/Funnel/Funnel';
 import Step from 'components/@common/Funnel/Step';
 
@@ -34,15 +31,11 @@ export default function StackNavigator({isLoggedIn}: StackNavigatorProps) {
             ? TabNavigator
             : () => (
                 <Funnel steps={steps} step={currentStep}>
-                  <Step name={Steps.Step1}>
-                    <Description />
-                  </Step>
-                  <Step name={Steps.Step2}>
-                    <Login />
-                  </Step>
-                  <Step name={Steps.Step3}>
-                    <UserInfo />
-                  </Step>
+                  {stepInfoList.map(({name, component}) => (
+                    <Step key={name} name={name}>
+                      {component}
+                    </Step>
+                  ))}
                 </Funnel>
               )
         }
