--- src/navigators/constants/step.ts
@@ -1,9 +1,0 @@
-import {NonEmptyArray} from 'types/funnel';
-
-export const steps: NonEmptyArray<string> = ['Step1', 'Step2', 'Step3'];
-
-export const enum Steps {
-  Step1 = 'Step1',
-  Step2 = 'Step2',
-  Step3 = 'Step3',
-}
