+++ src/navigators/types/steps.ts
@@ -1,0 +1,6 @@
+import {Steps} from 'navigators/constants/steps';
+
+export type StepInfo = {
+  name: Steps[number];
+  component: React.ReactElement;
+};
